Chore(messaging): move messages logic to service (#328)
* chore(messaging): move routes logic to services

* Updated sdks
